monitoramento_indiretos/pyproject.toml
[project]
name = "monitoramento-indiretos"
version = "0.1.0"
description = ""
authors = [
    {name = "zPaulo",email = "paulotjs147@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "python-jose (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
